Use connection.set instead of connection.append
Repeated cache puts with the same key were appending new data to
existing data, when they should have been replacing it. This
caused the JSON to be unparseable, meaning that the cache was
unusable - thus causing the runner to wait several hundred ms on
each request, rather than the 1 or 2 ms you'd expect for a cached
value.

This PR fixes the issue