Add Kubernetes readiness probe
Even though we are doing rolling deploys, it seems that traffic is being served
to the pods before Rails has booted. So this is still resulting in downtime.

By exposing an endpoint for the readiness probe to test, we can ensure that the application
will be ready to serve traffic by the time it is registered with the load balancer.