Notify slack on deployment failures and successes
Moving the acceptance tests into each deployment has removed the
notification that happens when they fail as it's not actually the
circleci job that is running anymore.

We require notifications to be sent to the #form-builder-deployments
channel on failure of acceptance tests.

Also on failure to deploy to production.

Also on a successful deploy to production